pivoting project to Pass Rush Predictor
removing previous files. Also copied the original code into the PassRushPredictor file, and then overhauled all the code to UpdatedPRP so it works with current versions of all the libraries used. Mainly it was changes to the way pytorch was interacting with pandas so a fair amount was changed. But it WORKS!